Add unfinished-ssg package with initial implementation and configuration
