+++ packages/unfinished-ssg/package.json
@@ -1,9 +1,24 @@
 {
   "name": "@adbl/unfinished-ssg",
+  "version": "0.0.0",
+  "description": "Static site generation for the unfinished UI framework.",
+  "private": false,
+  "type": "module",
+  "scripts": {
+    "build": "node scripts/build.js"
+  },
+  "dependencies": {
+    "domhandler": "^5.0.3",
+    "htmlparser2": "^10.0.0"
+  },
+  "exports": {
+    "./plugin": {
+      "import": "./dist/plugin.js",
+      "types": "./dist/plugin.d.ts"
+    }
+  },
   "devDependencies": {
     "typescript": "^5.8.2",
-    "domhandler": "^5.0.3",
-    "htmlparser2": "^10.0.0",
     "vite": "^5.4.14"
   },
   "peerDependencies": {
