+++ packages/unfinished-ssg/source/plugin.js
@@ -1,0 +1,5 @@
+export function unfinishedSSG() {
+  return {
+    name: 'vite-unfinished-ssg',
+  };
+}
